+++ packages/components/src/molecules/Accordion/AccordionItem.tsx
@@ -40,9 +40,10 @@ type AccordionItemComponent = <C extends ElementType = 'div'>(
   props: AccordionItemProps<C>
 ) => ReactElement | null
 
-const AccordionItem = forwardRef(function AccordionItem<
-  C extends ElementType = 'div',
->(
+const AccordionItem = forwardRef<
+  HTMLDivElement,
+  Omit<AccordionItemProps<'div'>, 'ref'>
+>(function AccordionItem<C extends ElementType = 'div'>(
   {
     prefixId = '',
     index = 0,