+++ packages/core/src/components/sections/Navbar/Navbar.tsx
@@ -4,6 +4,7 @@ import Section from '../Section'
 import styles from './section.module.scss'
 
 import Navbar from '../../../components/navigation/Navbar'
+import { DEFAULT_FILE_UPLOAD_CARD_PROPS } from '../../../components/search/fileUploadCardDefaults'
 import { NavbarDefaultComponents } from './DefaultComponents'
 
 type PageLinks = {
@@ -38,6 +39,22 @@ export interface NavbarProps {
           }
         }
       }
+    }
+    /** Props for FileUploadCard (labels, messages). Pass from CMS so all copy is editable. */
+    fileUploadCardProps?: {
+      title?: string
+      fileInputAriaLabel?: string
+      dropzoneAriaLabel?: string
+      dropzoneTitle?: string
+      selectFileButtonLabel?: string
+      downloadTemplateButtonLabel?: string
+      removeButtonAriaLabel?: string
+      searchButtonLabel?: string
+      uploadingStatusText?: string
+      getCompletedStatusText?: (fileSize: number) => string
+      errorMessages?: Partial<
+        Record<string, { title: string; description: string }>
+      >
     }
   }
   signInButton: {
@@ -96,7 +113,11 @@ function NavbarSection({
         home={home}
         menu={menu}
         logo={logo}
-        searchInput={searchInput}
+        searchInput={{
+          ...searchInput,
+          fileUploadCardProps:
+            searchInput.fileUploadCardProps ?? DEFAULT_FILE_UPLOAD_CARD_PROPS,
+        }}
         cart={cartIcon}
         links={pageLinks}
         signIn={{ button: signInButton }}